requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
idna==3.10
numpy==2.1.2
probableparsing==0.0.1
python-crfsuite==0.9.11
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
usaddress==0.5.11
